types: update storage types
