+++ types/src/common.ts
@@ -44,6 +44,9 @@ export function JoyBTreeSet<V extends UInt>(valType: Constructor<V>): Constructo
   }
 }
 
+export class ChannelId extends u64 {}
+export class DAOId extends u64 {}
+
 // common types between Forum and Proposal Discussions modules
 export class ThreadId extends u64 {}
 export class PostId extends u64 {}
@@ -146,6 +149,8 @@ export const commonTypes: RegistryTypes = {
   MemoText,
   Address,
   LookupSource,
+  ChannelId,
+  DAOId,
 }
 
 export default commonTypes
