+++ types/src/storage.ts
@@ -1,5 +1,5 @@
-import { Option, Vec as Vector, BTreeMap, u64, bool, Text, Null } from '@polkadot/types'
-import { BlockAndTime, JoyEnum, JoyStructDecorated, Hash } from './common'
+import { Option, Vec, BTreeMap, u64, bool, Text, Null, Bytes } from '@polkadot/types'
+import { BlockAndTime, JoyEnum, JoyStructDecorated, Hash, ChannelId, DAOId } from './common'
 import { MemberId } from './members'
 import { StorageProviderId } from './working-group' // this should be in discovery really
 import { randomAsU8a } from '@polkadot/util-crypto'
@@ -29,7 +29,7 @@ export class ContentId extends Hash {
 export class DataObjectTypeId extends u64 {}
 export class DataObjectStorageRelationshipId extends u64 {}
 
-export class VecContentId extends Vector.with(ContentId) {}
+export class VecContentId extends Vec.with(ContentId) {}
 export class OptionVecContentId extends Option.with(VecContentId) {}
 
 export const LiaisonJudgementDef = {
@@ -40,8 +40,40 @@ export const LiaisonJudgementDef = {
 export type LiaisonJudgementKey = keyof typeof LiaisonJudgementDef
 export class LiaisonJudgement extends JoyEnum(LiaisonJudgementDef) {}
 
+export class WorkingGroupType extends JoyEnum({
+  ContentDirectory: Null,
+  Builders: Null,
+  StorageProviders: Null,
+}) {}
+
+export class AbstractStorageObjectOwner extends JoyEnum({
+  Channel: ChannelId,
+  DAO: DAOId,
+  Council: Null,
+  WorkingGroup: WorkingGroupType,
+}) {}
+
+export class StorageObjectOwner extends JoyEnum({
+  Member: MemberId,
+  AbstractStorageObjectOwner: AbstractStorageObjectOwner,
+}) {}
+
+export class ContentParameters extends JoyStructDecorated({
+  content_id: ContentId,
+  type_id: DataObjectTypeId,
+  size: u64,
+  ipfs_content_id: Bytes,
+}) {
+  /** Actually it's 'size', but 'size' is already reserved by a parent class. */
+  get size_in_bytes(): u64 {
+    return this.get('size') as u64
+  }
+}
+
+export class Content extends Vec.with(ContentParameters) {}
+
 export class DataObject extends JoyStructDecorated({
-  owner: MemberId,
+  owner: StorageObjectOwner,
   added_at: BlockAndTime,
   type_id: DataObjectTypeId,
   size: u64,
@@ -77,6 +109,11 @@ export const mediaTypes: RegistryTypes = {
   DataObjectTypeId,
   DataObjectType,
   DataObjectsMap,
+  ContentParameters,
+  StorageObjectOwner,
+  AbstractStorageObjectOwner,
+  WorkingGroupType,
+  Content,
 }
 
 export default mediaTypes
